Clean up import controller by extracting logic do dedicated import session service
